requirements.txt
antlr4-python3-runtime==4.9.3
cachetools==5.3.3
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
h5pickle==0.4.2
h5py==3.11.0
hdf5plugin==4.4.0
hydra-core==1.3.2
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sympy==1.12
tbb==2021.12.0
tqdm==4.66.4
typing_extensions==4.11.0
